agent_ai/src/requirements.txt
azure-search-documents
azure-identity
azure-ai-formrecognizer
azure-cosmos

fastapi
uvicorn
python-dotenv
python-multipart
pytz
ipython

openai
langchain
langchain-core
langchain-community
langchain-openai
langgraph
azure-storage-file-datalake
azure-identity
msal

scrapegraphai
#==1.37.0
nest-asyncio==1.6.0

pandas==2.2.3
openpyxl==3.1.5
python-docx ==1.1.2
playwright


#wa
python-socketio
aiohttp
asyncio
python-engineio 